Split up stuff using include
